Alert mapper code added

+++ dotnet/corecomponents/net.atos.daf.ct2.alert/entity/AlertFilterRef.cs
@@ -14,7 +14,7 @@ namespace net.atos.daf.ct2.alert.entity
 
         public string FilterType { get; set; }
 
-        public decimal ThresholdValue { get; set; }
+        public double ThresholdValue { get; set; }
 
         public string UnitType { get; set; }
